pyproject.toml
[project]
name = "auto-dotenv"
version = "0.1.0"
description = "Automatically load .env variables if auto-dotenv is installed"
keywords=['environment variables', 'configurations', 'secrets', 'env', 'dotenv']
authors = [
    { name = "Dhia Hmila", email = "dhiahmila.dev@gmail.com" }
]
maintainers = [{ name = "Dhia Hmila", email = "dhiahmila.dev@gmail.com" }]
dependencies = [
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
]

[project.urls]
Repository = "https://github.com/hmiladhia/auto-dotenv"
Issues = "https://github.com/hmiladhia/auto-dotenv/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel.hooks.autorun]
dependencies = ["hatch-autorun"]
code = "__import__('auto_dotenv').autoload_dotenv()"

[tool.hatch.build.targets.wheel]
only-include=["src/auto_dotenv.py"]
sources=["src"]

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.0.2",
]

[tool.ruff]
line-length = 88
target-version = "py38"
include = ["*.py", "*.pyi", "**/pyproject.toml", "*.ipynb"]


[tool.ruff.lint.pycodestyle]
max-line-length = 100
